pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthetic-experiments"
version = "0.1.0"
description = "A framework for social scientists to conduct experimental research on LLM conversations"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Alex Newhouse"}
]
keywords = ["llm", "social-science", "experiments", "research", "conversations", "chatbots"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "ollama>=0.1.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "pandas>=2.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
analysis = [
    "textblob>=0.17.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
synthetic-exp = "synthetic_experiments.cli.main:cli"

[project.urls]
Homepage = "https://github.com/alexnewhouse/synthetic-experiment-framework"
Issues = "https://github.com/alexnewhouse/synthetic-experiment-framework/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["synthetic_experiments*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
